Fix current tests structure
