+++ Liviano.Tests/Liviano/Functional/WalletExtensionsTest.cs
@@ -1,21 +1,46 @@
+//
+// BlockcypherBlock.cs
+//
+// Author:
+//       igor <igorgue@protonmail.com>
+//
+// Copyright (c) 2022 HODL Wallet
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
 using System;
+using System.IO;
 using System.Net;
 using System.Linq;
 using System.Net.Http;
 using System.Collections.Generic;
 using System.Threading;
-using Xunit;
-
-using Newtonsoft.Json;
+using System.Threading.Tasks;
 
 using NBitcoin;
 using NBitcoin.DataEncoders;
-using System.IO;
-using System.Threading.Tasks;
+using Newtonsoft.Json;
+using Xunit;
 using Xunit.Abstractions;
+
 using Liviano.Extensions;
 
-namespace Liviano.Tests.Liviano
+namespace Liviano.Tests.Funcional
 {
     class BlockcypherBlock
     {
